+++ src/err.h
@@ -5,10 +5,10 @@ typedef enum { LERR_DIV_ZERO, LERR_BAD_NUM, LERR_BAD_OP } LERR_t;
 
 typedef struct {
   LVAL_e type;
-  long num;
+  double num;
   LERR_t err;
 } lval_t;
 
-lval_t lval_eval(long x, LVAL_e type);
+lval_t lval_eval(double x, LVAL_e type);
 void lval_print(lval_t v);
 void lval_println(lval_t v);
